// src/app/views/production-planning/interfaces/inventory-item-status-interface.ts
export interface InventoryItemStatusEntity {
    InventoryItemId: number;
    ItemStatus?: string;
    ItemColor?: string;
    ItemPriority: number;
    TenantId: number;
    CreatedBy: number;
    ModifiedBy?: number;
}

export interface InventoryItemStatusSelectEntity {
    InventoryItemId: number;
    TenantId: number;
}

export interface InventoryItemStatusDeleteEntity {
    InventoryItemId: number;
}